feat: Add plant database view
Introduces a new view for browsing plant species and allows users to search for information. This includes a new `database` view state, the `SearchIcon` and `SunIcon` components, and a `Species` type definition. The navigation bar has been updated to include a button for the new database view. A mock `SPECIES_DB` is also added for initial data.